implemented user functionalities and connected with remote database
