+++ user-management-service/config.env
@@ -20,3 +20,9 @@ EMAIL_PASSWORD_DEV=7f3b70db5436ee
 EMAIL_PROD=smartpossystemuom@gmail.com
 PASSWORD_PROD=lsex sdom yvsy mecr
 
+DB_TYPE=postgres
+PG_HOST=db
+PG_PORT=5432
+PG_USER=postgres
+PG_PASSWORD=postgres
+PG_DB=postgres
