+++ user-management-service/src/Controllers/authController.ts
@@ -10,47 +10,16 @@ import { promisify } from "util";
 import crypto from "crypto";
 import { sendMail } from "../Utils/userMail";
 import { Employee } from "../entity/Employee";
+import {
+  changedPasswordAfter,
+  correctPassword,
+  createPasswordResetToken,
+  createSendToken,
+  isEligible,
+  verifyToken,
+} from "../Services/authServices";
 
 dotenv.config({ path: "../../config.env" });
-
-const signToken = (email: string): string => {
-  const jwtSecret = process.env.JWT_SECRET as string;
-  const jwtExpiresIn = process.env.JWT_EXPIRES_IN as string;
-
-  return jwt.sign({ email }, jwtSecret, {
-    expiresIn: jwtExpiresIn,
-  });
-};
-
-const createSendToken = (
-  user: any,
-  statusCode: number,
-  res: Response
-): void => {
-  const token = signToken(user.email);
-
-  const jwtCookieExpiresIn = Number(process.env.JWT_COOKIE_EXPIRES_IN_HOURS);
-
-  const cookieOptions: { expires: Date; httpOnly: boolean; secure?: boolean } =
-    {
-      expires: new Date(Date.now() + jwtCookieExpiresIn * 3600 * 1000),
-      httpOnly: true,
-    };
-
-  if (process.env.NODE_ENV === "production") {
-    cookieOptions.secure = true;
-  }
-
-  res.cookie("jwt", token, cookieOptions);
-
-  res.status(statusCode).json({
-    status: "success",
-    data: {
-      //  token,
-      user: { ...user, password: undefined },
-    },
-  });
-};
 
 export const login = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
@@ -74,25 +43,6 @@ export const login = catchAsync(
     createSendToken(user, 200, res);
   }
 );
-
-export const correctPassword = async function (
-  candidatePassword: any,
-  userPassword: any
-) {
-  return await bcrypt.compare(candidatePassword, userPassword);
-};
-
-const verifyToken = (token: string, secret: string): Promise<any> => {
-  return new Promise((resolve, reject) => {
-    jwt.verify(token, secret, (err: any | null, decoded: any) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(decoded);
-      }
-    });
-  });
-};
 
 export const protect = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
@@ -157,53 +107,6 @@ export const protect = catchAsync(
   }
 );
 
-const changedPasswordAfter = function (
-  JWTTimestamp: number,
-  passwordChangedAt: Date | null
-) {
-  if (passwordChangedAt) {
-    const changedTimestamp = passwordChangedAt.getTime() / 1000;
-    return JWTTimestamp < changedTimestamp;
-  }
-  // Assuming passwordChangedAt is null
-  return false;
-};
-
-const isEligible = (userRole: string, role: string): boolean => {
-  const roles1 = [
-    "General Manager",
-    "Regional Manager",
-    "Store Manager",
-    "Store Supervisor",
-    "Cashier",
-  ];
-  const roles2 = [
-    "General Manager",
-    "Regional Manager",
-    "Inventory Manager",
-    "Inventory Supervisor",
-  ];
-
-  const roles =
-    roles1.includes(userRole) && roles1.includes(role)
-      ? roles1
-      : roles2.includes(userRole) && roles2.includes(role)
-      ? roles2
-      : null;
-
-  if (roles === null) {
-    return false;
-  }
-
-  const userRoleIndex = roles.indexOf(userRole);
-  const roleIndex = roles.indexOf(role);
-
-  if (userRoleIndex < 0 || roleIndex < 0) {
-    return false;
-  }
-  return userRoleIndex < roleIndex;
-};
-
 export const restrictToCreate = (
   req: Request,
   res: Response,
@@ -241,19 +144,6 @@ export const restrictTo = (...roles: any) => {
 
     next();
   };
-};
-
-const createPasswordResetToken = function (user: any) {
-  const resetToken = crypto.randomBytes(32).toString("hex");
-
-  user.password_reset_token = crypto
-    .createHash("sha256")
-    .update(resetToken)
-    .digest("hex");
-
-  user.password_reset_expires = new Date(Date.now() + 10 * 60 * 1000);
-
-  return resetToken;
 };
 
 export const forgotPassword = catchAsync(
@@ -347,6 +237,7 @@ export const resetPassword = catchAsync(
     createSendToken(user, 200, res);
   }
 );
+
 export const logout = (req: Request, res: Response) => {
   res.cookie("jwt", "loggedout", {
     expires: new Date(Date.now() + 10 * 1000),
