+++ user-management-service/src/Controllers/employeeController.ts
@@ -1,93 +1,11 @@
 import { NextFunction, Request, Response } from "express";
-import { User } from "../Utils/database";
 import bcrypt from "bcryptjs";
-import { body, validationResult } from "express-validator";
-import { validRoles } from "../Utils/roles";
-import { sendMail } from "../Utils/userMail";
 import AppError from "../Utils/appError";
 import catchAsync from "../Utils/catchAsync";
-import { AppDataSource } from "./../index";
+import { AppDataSource } from "../index";
 import { Employee } from "../entity/Employee";
 import { Role } from "../entity/Role";
-import {
-  lowerRole1,
-  lowerRole2,
-  roleHierarchy1,
-  roleHierarchy2,
-  roleHierarchy3,
-} from "../Utils/roleHierarchy";
-
-export const isUserExists = catchAsync(
-  async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const userRepository = AppDataSource.getRepository(Employee);
-      const { email } = req.body;
-      const user = await userRepository.findOne({
-        where: {
-          email: email,
-        },
-      });
-      if (user) {
-        next(new AppError("User already exists", 404));
-      }
-      next();
-    } catch (err: any) {
-      return next(new AppError(err.message, 400));
-    }
-  }
-);
-
-export const validateUser = [
-  body("name")
-    .isString()
-    .withMessage("User's name must be a string")
-    .notEmpty()
-    .withMessage("User's name is required"),
-  body("role_name")
-    .isIn(validRoles)
-    .withMessage(
-      "Role must be one of Regional Manager, Inventory Manager, Inventory Supervisor, Store Manager, Store Supervisor, Cashier"
-    )
-    .notEmpty()
-    .withMessage("Role is required"),
-  body("email")
-    .isEmail()
-    .withMessage("Email must be a valid email address")
-    .notEmpty()
-    .withMessage("Email is required"),
-  catchAsync(async (req: Request, res: Response, next: NextFunction) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return next(new AppError(errors.array()[0].msg, 400));
-    }
-    next();
-  }),
-];
-
-export const sendMailToUser = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  const urlForUser = `${req.protocol}://${req.get("host")}/api/v1/createUser`;
-  const message = `You have been added to our POS System as a ${req.body.role_name}. Please click on the link below to set your username and password to complete your account. \n\n ${urlForUser} \n\n If you did not request this, please ignore this email.`;
-
-  try {
-    await sendMail({
-      email: req.body.email,
-      subject: "Complete your registration!",
-      message,
-    });
-  } catch (err: any) {
-    return next(
-      new AppError(
-        "There was an error sending the email. Try again later!",
-        500
-      )
-    );
-  }
-  next();
-};
+import { getCurrentUserRoleInfo } from "../Services/emplyeeServices";
 
 export const createUserByAdmin = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
@@ -131,31 +49,6 @@ export const createUserByAdmin = catchAsync(
     }
   }
 );
-
-export const validateCreation = [
-  body("password")
-    .isString()
-    .withMessage("Password must be a string")
-    .notEmpty()
-    .withMessage("Password is required"),
-  body("passwordConfirm")
-    .isString()
-    .withMessage("Password Confirmation must be a string")
-    .notEmpty()
-    .withMessage("Password Confirmation is required"),
-  body("email")
-    .isEmail()
-    .withMessage("Email must be a valid email address")
-    .notEmpty()
-    .withMessage("Email is required"),
-  catchAsync(async (req: Request, res: Response, next: NextFunction) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return next(new AppError(errors.array()[0].msg, 400));
-    }
-    next();
-  }),
-];
 
 export const createUserByUser = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
@@ -269,27 +162,50 @@ export const getOneUser = catchAsync(
   }
 );
 
-const getCurrentUserRoleInfo = (req: Request) => {
-  if (!req.user || !req.user.role || !req.user.role.role_name) {
-    throw new Error("There is a problem with the logged-in user");
+/*export const createAdmin = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { name, role_name, password, passwordConfirm, email } = req.body;
+
+    if (password.startsWith("tempPassword")) {
+      return next(new AppError("Please set another word as the password", 400));
+    }
+
+    if (password !== passwordConfirm) {
+      return next(new AppError("Passwords do not match", 400));
+    }
+
+    if (password.length <= 8) {
+      return next(new AppError("Password must be at least 8 characters", 400));
+    }
+
+    const userRepository = AppDataSource.getRepository(Employee);
+    const roleRepository = AppDataSource.getRepository(Role);
+
+    const role = await roleRepository.findOneBy({ role_name });
+
+    if (!role) {
+      return next(new AppError("Role not found", 400));
+    }
+
+    const hashedPassword = await bcrypt.hash(password, 12);
+
+    const newUser = userRepository.create({
+      employee_name: name,
+      email,
+      password: hashedPassword,
+      role,
+      temporary: false,
+      is_active: true,
+    });
+
+    await userRepository.save(newUser);
+
+    res.status(201).json({
+      status: "success",
+      data: {
+        user: { ...newUser, password: undefined },
+        message: "Account created successfully",
+      },
+    });
   }
-
-  const currentUserRole = req.user.role.role_name;
-
-  let roleHierarchy: { [key: string]: number };
-
-  if (lowerRole1.includes(currentUserRole)) {
-    roleHierarchy = roleHierarchy2;
-  } else if (lowerRole2.includes(currentUserRole)) {
-    roleHierarchy = roleHierarchy3;
-  } else {
-    roleHierarchy = roleHierarchy1;
-  }
-
-  const currentUserRoleLevel = roleHierarchy[currentUserRole];
-  const allowedRoles = Object.entries(roleHierarchy)
-    .filter(([_, level]) => level >= currentUserRoleLevel)
-    .map(([role]) => role);
-
-  return allowedRoles;
-};
+); */
