+++ user-management-service/src/Controllers/errorController.ts
@@ -21,6 +21,7 @@ const sendErrorProd = (err: any, res: Response) => {
   if (err.isOperational) {
     res.status(err.statusCode).json({
       status: err.status,
+      statusCode: err.statusCode,
       message: err.message,
     });
