+++ user-management-service/src/app.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { userRouter } from "./Routes/userRoutes";
+import { userRouter } from "./Routes/employeeRoutes";
 import cookieParser from "cookie-parser";
 import rateLimit from "express-rate-limit";
