+++ user-management-service/src/index.ts
@@ -4,6 +4,8 @@ import app from "./app";
 import { Employee } from "./entity/Employee";
 import { Customer } from "./entity/Customer";
 import { Role } from "./entity/Role";
+//import { insertRoles } from "./tests/insertRoles";
+//import { insertEmployees } from "./tests/insertEmployees";
 
 process.on("uncaughtException", (err: Error) => {
   console.log("UNCAUGHT EXCEPTION! 💥 Shutting down...");
@@ -13,13 +15,27 @@ process.on("uncaughtException", (err: Error) => {
 
 dotenv.config({ path: "./config.env" });
 
-export const AppDataSource = new DataSource({
+//For local database
+/* export const AppDataSource = new DataSource({
   type: process.env.DB_TYPE as any,
   host: process.env.DB_HOST,
   port: parseInt(process.env.DB_PORT || "", 10),
   username: process.env.DB_USER,
   password: process.env.DB_PASSWORD,
   database: process.env.DB_NAME,
+  synchronize: true,
+  logging: false,
+  entities: [Employee, Customer, Role],
+}); */
+
+//For remote database
+export const AppDataSource = new DataSource({
+  type: process.env.DB_TYPE as any,
+  host: process.env.PG_HOST,
+  port: parseInt(process.env.PG_PORT || "", 10),
+  username: process.env.PG_USER,
+  password: process.env.PG_PASSWORD,
+  database: process.env.PG_DB,
   synchronize: true,
   logging: false,
   entities: [Employee, Customer, Role],
