+++ user-management-service/src/tests/insertRoles.ts
@@ -1,4 +1,4 @@
-/*import { AppDataSource } from "../data-source";
+import { AppDataSource } from "../data-source";
 import { Role } from "../entity/Role";
 
 export async function insertRoles() {
@@ -18,4 +18,3 @@ export async function insertRoles() {
     );
   }
 }
-*/
